back_end/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
pymongo==4.6.1

# AI and ML
openai==1.12.0
numpy==1.26.3
tiktoken==0.5.2

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0

# Environment and utilities
python-dotenv==1.0.0
pydantic==2.5.3
